Created the GroovyPoliceBot
Created a functional bot which polices where you can post GroovyBot commands in a discord server.

Deletes groovy command messages if they are placed in the wrong channel
Prints a message telling the person off.
Adds a strike to the person. 3 strikes means a mild, mostly annoying punishment. This is yet to be implemented.